Moved subclass-specific code out of PlotPanel and into their appropriate subclasses, e.g. FacsPlotPanel or CenterChooserDialog. Specifically, figure click code and setting current subplot.
Added method to load saved subplots.

Subplots can now be linked to the current axes selection combo boxes or unlinked, ie when the user changes axes selection, the subplot does not change.

All parameters to Subplot constructor have defaults so that a Subplot instance can be created without passing any parameters. This is done so that the new load() method can specify all the parameters from a dictionary saved to the system state save file.

Added data access properties: Data, Clustering, and Labels to Subplot.

Added code to FacsTreeCtrlPanel to effect data tree item renaming, and reflect the movement of the clusteringInfo() method.