Added a checkbox to the axis selector panel. This allows users to determine whether subplots will respond to changes in the axis selection. The checkbox is linked to the currently selected subplot. It displays the current status, and clicking it changes the status. If checked, the subplot will be updated with the change in axes. If unchecked, the subplot display will remain static, displaying the axes that were selected when it was unchecked.
Added menu support for saving and loading the system state.

Used the correct wx IDs for the Exit and About menus so they show up in the appropriate menu for their platform.

Moved the code for setting the contents of the axes selection combo boxes to its own method.

Created two exit handling methods. One for a window-closing event (which calls the next), and a program exit event. Note that wxWidgets provides events for OS logoff and shutdown events, allowing you to hook in and potentially prevent either, or at least ask the user to save their work before logoff/shutdown proceeds.

Renamed all event parameters in the handler methods from 'e' to 'event'.

All dialogs in this module are now destroyed after use. Not doing this prevents the program from exiting cleanly (ie without having to kill it in unpleasant ways).

If the error reporting was unsuccessful, the user is now notified via a message in the status bar.

The program now exits appropriately (with system.exit()) when the main window is closed.